Converts RNDTS quantity to float

+++ transform/models/trusted_zone/rndts/dnd_entrants.sql
@@ -42,7 +42,7 @@ select
     raison_sociale_declarant,
     code_dechet,
     denomination_usuelle,
-    quantite,
+    quantite::numeric,
     unite,
     date_reception::date,
     heure_pesee::time,
